#158 integration of jxbrowser 7.0 using new APIs

+++ eclipse-plugin-studio/src/com/twinsoft/convertigo/eclipse/swt/C8oBrowser.java
@@ -22,6 +22,8 @@ package com.twinsoft.convertigo.eclipse.swt;
 import java.awt.Frame;
 import java.io.File;
 import java.io.IOException;
+import java.net.ServerSocket;
+import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -32,37 +34,28 @@ import org.eclipse.swt.events.DisposeEvent;
 import org.eclipse.swt.events.DisposeListener;
 import org.eclipse.swt.widgets.Composite;
 
-import com.teamdev.jxbrowser.chromium.Browser;
-import com.teamdev.jxbrowser.chromium.BrowserContext;
-import com.teamdev.jxbrowser.chromium.BrowserContextParams;
-import com.teamdev.jxbrowser.chromium.BrowserException;
-import com.teamdev.jxbrowser.chromium.BrowserPreferences;
-import com.teamdev.jxbrowser.chromium.events.FailLoadingEvent;
-import com.teamdev.jxbrowser.chromium.events.FinishLoadingEvent;
-import com.teamdev.jxbrowser.chromium.events.FrameLoadEvent;
-import com.teamdev.jxbrowser.chromium.events.LoadEvent;
-import com.teamdev.jxbrowser.chromium.events.LoadListener;
-import com.teamdev.jxbrowser.chromium.events.ProvisionalLoadingEvent;
-import com.teamdev.jxbrowser.chromium.events.StartLoadingEvent;
-import com.teamdev.jxbrowser.chromium.swing.BrowserView;
+import com.teamdev.jxbrowser.browser.Browser;
+import com.teamdev.jxbrowser.engine.Engine;
+import com.teamdev.jxbrowser.engine.EngineOptions;
+import com.teamdev.jxbrowser.engine.RenderingMode;
+import com.teamdev.jxbrowser.navigation.event.LoadFinished;
+import com.teamdev.jxbrowser.view.swing.BrowserView;
 import com.twinsoft.convertigo.beans.core.Project;
-import com.twinsoft.convertigo.engine.Engine;
+import com.twinsoft.convertigo.engine.EnginePropertiesManager;
+import com.twinsoft.convertigo.engine.util.Crypto2;
 import com.twinsoft.convertigo.engine.util.FileUtils;
 
 public class C8oBrowser extends Composite {
 	
-	static {
-		int port = 18082;
-		BrowserPreferences.setChromiumSwitches("--remote-debugging-port=" + port);
-	}
-	
 	private static Thread threadSwt = null;
-	private static Map<String, BrowserContext> browserContexts = new HashMap<>();
+	private static Map<String, Engine> browserContexts = new HashMap<>();
+	private static final String jxKey = "x31d140170e500d2cb6ea40186207fc4648722a44e93ce99816d812fb04ad802339785879b7429b466dea337c536d1d13aa4de4d2c578c252dd5885460c0f511067527743a0df65e6be52632fe108f782";
 
+	private String debugUrl;
 	private BrowserView browserView;
 
-	private void init(Composite parent, BrowserContext browserContext) {
-		browserView = new BrowserView(new Browser(browserContext));
+	private void init(Composite parent, Engine browserContext) {
+		browserView = BrowserView.newInstance(browserContext.newBrowser());
 		Frame frame = SWT_AWT.new_Frame(this);
 		frame.add(browserView);
 		threadSwt = parent.getDisplay().getThread();
@@ -76,36 +69,53 @@ public class C8oBrowser extends Composite {
 	}
 	
 	public C8oBrowser(Composite parent, int style) {
-		super(parent, style | SWT.EMBEDDED | SWT.NO_BACKGROUND);
-		init(parent, BrowserContext.defaultContext());
+		this(parent, style, (Project) null);
 	}
 
 	public C8oBrowser(Composite parent, int style, Project project) {
 		super(parent, style | SWT.EMBEDDED | SWT.NO_BACKGROUND);
 		boolean retry = false;
 		do {
-			File browserIdFile = new File(project.getDirPath() + "/_private/browser_id");
-			String browserId = Long.toString(System.currentTimeMillis(), Character.MAX_RADIX);
-			try {
-				browserId = FileUtils.readFileToString(browserIdFile, "UTF-8");
-			} catch (Exception e) {
+			File browserIdFile = null;
+			String browserId = "default";
+			if (project != null) {
+				browserIdFile = new File(project.getDirPath() + "/_private/browser_id");
+				browserId = Long.toString(System.currentTimeMillis(), Character.MAX_RADIX);
 				try {
-					FileUtils.write(browserIdFile, browserId, "UTF-8");
-				} catch (IOException e1) {
-				}
+					browserId = FileUtils.readFileToString(browserIdFile, "UTF-8");
+				} catch (Exception e) {
+					try {
+						FileUtils.write(browserIdFile, browserId, "UTF-8");
+					} catch (IOException e1) {
+					}
+				}	
 			}
-			File browserWorks = new File(Engine.USER_WORKSPACE_PATH + "/browser-works");
+			
+			File browserWorks = new File(com.twinsoft.convertigo.engine.Engine.USER_WORKSPACE_PATH + "/browser-works");
 			browserWorks.mkdirs();
-			BrowserContext browserContext = browserContexts.get(browserId);
+			Engine browserContext = browserContexts.get(browserId);
 			if (browserContext == null) {
-				browserContext = new BrowserContext(new BrowserContextParams(Engine.USER_WORKSPACE_PATH + "/browser-works/" + browserId));
+				int debugPort; 
+				try (ServerSocket sock = new ServerSocket(0)) {
+					debugPort = sock.getLocalPort();
+				} catch (Exception e) {
+					debugPort = 18081 + browserContexts.size();
+				}
+				String key = Crypto2.decodeFromHexString(EnginePropertiesManager.PropertyName.CRYPTO_PASSPHRASE.getDefaultValue(), jxKey);
+				browserContext = Engine.newInstance(EngineOptions.newBuilder(RenderingMode.HARDWARE_ACCELERATED)
+						.userDataDir(Paths.get(com.twinsoft.convertigo.engine.Engine.USER_WORKSPACE_PATH, "browser-works", browserId))
+						.licenseKey(key)
+						.remoteDebuggingPort(debugPort).build());
 				browserContexts.put(browserId, browserContext);
 			}
+			debugUrl = "http://localhost:" + browserContext.options().remoteDebuggingPort().get();
 			try {
 				init(parent, browserContext);
-			} catch (BrowserException e) {
+			} catch (Exception e) {
 				if (!retry) {
-					browserIdFile.delete();
+					if (browserIdFile != null) {
+						browserIdFile.delete();
+					}
 					retry = true;
 				} else {
 					throw e;
@@ -114,7 +124,7 @@ public class C8oBrowser extends Composite {
 		} while (retry);
 	}
 
-	public C8oBrowser(Composite parent, int style, BrowserContext browserContext) {
+	public C8oBrowser(Composite parent, int style, Engine browserContext) {
 		super(parent, style | SWT.EMBEDDED | SWT.NO_BACKGROUND);
 		init(parent, browserContext);
 	}
@@ -122,7 +132,7 @@ public class C8oBrowser extends Composite {
 	@Override
 	public void dispose() {
 		run(() -> {
-			getBrowser().dispose();
+			getBrowser().close();;
 		});
 		super.dispose();
 	}
@@ -143,20 +153,20 @@ public class C8oBrowser extends Composite {
 			String link = bg.getRed() < 128 ? "cyan" : "blue";
 			html = html.replace("$background$", background).replace("$foreground$", foreground).replace("$link$", link);
 		}
+		final String h = html;
 		if (html.contains("</html>")) {
-			getBrowser().loadHTML(html);
+			getBrowser().mainFrame().get().loadHtml(h);
 		} else {
-			getBrowser().getDocument().getDocumentElement().setInnerHTML(html);
+			getBrowser().mainFrame().get().document().get().documentElement().get().innerHtml(h);
 		}
 	}
 	
 	public void reloadText() {
-		Browser browser = getBrowser();
-		setText(browser.getHTML());
+		getBrowser().mainFrame().ifPresent(frame -> setText(frame.html()));
 	}
 
 	public void setUrl(String url) {
-		getBrowser().loadURL(url);
+		getBrowser().navigation().loadUrl(url);
 	}
 		
 	@Override
@@ -166,42 +176,24 @@ public class C8oBrowser extends Composite {
 	}
 
 	public void addProgressListener(ProgressListener progressListener) {
-		getBrowser().addLoadListener(new LoadListener() {
-			
-			@Override
-			public void onStartLoadingFrame(StartLoadingEvent event) {
-			}
-			
-			@Override
-			public void onProvisionalLoadingFrame(ProvisionalLoadingEvent event) {
-			}
-			
-			@Override
-			public void onFinishLoadingFrame(FinishLoadingEvent event) {
-				progressListener.completed(null);
-			}
-			
-			@Override
-			public void onFailLoadingFrame(FailLoadingEvent event) {
-				
-			}
-			
-			@Override
-			public void onDocumentLoadedInMainFrame(LoadEvent event) {
-			}
-			
-			@Override
-			public void onDocumentLoadedInFrame(FrameLoadEvent event) {
-				
-			}
+		getBrowser().navigation().on(LoadFinished.class, event -> {
+			progressListener.completed(null);
 		});
+	}
+	
+	public String getDebugUrl() {
+		return debugUrl;
 	}
 	
 	public static void run(Runnable runnable) {
 		if (threadSwt != null && threadSwt.equals(Thread.currentThread())) {
-			Engine.execute(runnable);
+			com.twinsoft.convertigo.engine.Engine.execute(runnable);
 		} else {
 			runnable.run();
 		}
+	}
+	
+	public static void main(String[] args) {
+		System.out.println(Crypto2.encodeToHexString(com.twinsoft.convertigo.engine.EnginePropertiesManager.PropertyName.CRYPTO_PASSPHRASE.getDefaultValue(), args[0]));
 	}
 }
